Authentication_and_Authorization/requirements.txt
Django==5.2
python-dotenv==1.1.0
PyJWT==2.10.1